Fixing Unit Tests & Code
Bug defect found in e2e test not caught in unit tests, issue lied in an internal function of calculate reuse distance in the popAndPushMemoryAddressBackOntoStack function not being passed in the distinctLocations vector. Unit test to verify is verifyNormalReuseHistogram, but it requires knowing the correct answer externally outside the code.